backend/requirements.txt
asgiref==3.5.2
Django==3.2.15
pytz==2022.2.1
sqlparse==0.4.2
psycopg2>=2.8
djangorestframework>=3.4.4